requirements.txt
Django>=5.1.3
pulsar-client==3.4.0
requests
docker
pyOpenSSL
django-extensions
Werkzeug
openpyxl
pandas
safety
defusedxml
whitenoise
FastAPI
uvicorn
python-dotenv
pydantic
email-validator
python-decouple 
mod_wsgi